refactor(client): change store file structure
