+++ packages/client/src/components/navigationBar/navigationBar.tsx
@@ -5,7 +5,7 @@ import { getUser } from '../../store/selectors'
 import avatarImg from '../../assets/images/avatar.png'
 
 import styles from './navigationBar.module.css'
-import { logout } from '../../store/actions/userActions'
+import { logout } from '../../store/user/userActions'
 
 export const NavigationBar = () => {
   const dispatch = useAppDispatch()
