+++ packages/client/src/pages/game/game.snap.test.tsx
@@ -1,7 +1,7 @@
 import * as renderer from 'react-test-renderer'
 import { Game } from './game'
 import configureStore from 'redux-mock-store'
-import { GameStatus } from '../../store/reducers/gameSlice'
+import { GameStatus } from '../../store/game/gameSlice'
 import { Provider } from 'react-redux'
 import { MemoryRouter } from 'react-router-dom'
