+++ packages/client/src/pages/leaderboard/leaderboard.tsx
@@ -13,8 +13,8 @@ import heroImg from '../../assets/images/hero.png'
 import baseStyles from '../../app/app.module.css'
 import styles from './leaderboard.module.css'
 import { Button } from '../../components/button/button'
-import { nextPage, previousPage } from '../../store/reducers/leaderboardSlice'
-import { leaderboardThunks } from '../../store/actions/leaderboardThunks'
+import { nextPage, previousPage } from '../../store/leadeboard/leaderboardSlice'
+import { leaderboardThunks } from '../../store/leadeboard/leaderboardThunks'
 import { useEffect } from 'react'
 import { ActivityIndicator } from '../../components/activityIndicator/activityIndicator'
